Move Glue.Period into data-lake
