+++ packages/punchcard/lib/storage/glue/index.ts
@@ -1,2 +1,1 @@
-export * from './period';
 export * from './table';
